requirements.txt
langchain 
langchain-core 
langchain-community
langchain-groq 
langchain-google-genai
streamlit 
python-dotenv 
fastapi 
uvicorn 